service-1/requirements.txt
fastapi
uvicorn
boto3
python-jose[cryptography]
